[Fragment] Move feature initialization into a separate class.
