+++ formula-android/src/main/java/com/instacart/formula/android/FragmentEnvironment.kt
@@ -1,8 +1,5 @@
 package com.instacart.formula.android
 
-import android.view.LayoutInflater
-import android.view.ViewGroup
-
 data class FragmentEnvironment(
     val logger: (String) -> Unit = {},
     val onScreenError: (FragmentKey, Throwable) -> Unit = { _, it -> throw it },
